hireZap/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.1
attrs==25.4.0
autobahn==25.10.1
Automat==25.4.16
av==16.0.1
billiard==4.2.4
boto3==1.42.21
botocore==1.42.21
cachetools==5.5.2
celery==5.6.1
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.7
ctranslate2==4.6.3
daphne==4.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.5
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-oauth-toolkit==1.7.1
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
faster-whisper==1.2.1
filelock==3.20.2
flatbuffers==25.12.19
fsspec==2025.12.0
gevent==25.9.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.6.2
mpmath==1.3.0
msgpack==1.1.2
numpy==2.4.0
oauthlib==3.3.1
onnxruntime==1.23.2
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
PyPDF2==3.0.1
pyreadline3==3.5.4
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.3
redis==6.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.16.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
zope.event==6.1
zope.interface==8.0.1
